change for the largest data type

+++ fi.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int fibSum (int rabbit);
-int fibNumber (int n);
+unsigned long long int fibSum (unsigned long long int rabbit);
+unsigned long long int fibNumber (int n);
 
 int main(){
     
@@ -16,13 +16,13 @@ int main(){
     }while(number<0);
 
     printf("\nYou entered %d\n", number);
-    printf("The sum of the first %d Fibonacci numbers is: %d\n\n",number, fibSum(number));
+    printf("The sum of the first %d Fibonacci numbers is: %llu\n\n",number, fibSum(number));
      
     
     return 0;
 }
 
-int fibNumber(int n){ 
+unsigned long long int fibNumber(int n){ 
     if(n==0){
         return 0;
     }else if(n==1){
@@ -33,9 +33,9 @@ int fibNumber(int n){
     }
 } 
 
-int fibSum (int rabbit){    
+unsigned long long int fibSum (unsigned long long int rabbit){    
     
-    int sum = 0;
+    unsigned long long int sum = 0;
     for(int i = 0; i<rabbit; i++){
          
         sum += fibNumber(i);
